requirements.txt
flask
flask_cors
flask_sqlalchemy
redis
mysqlclient
gunicorn
tencentcloud-sdk-python